Don't locate the bookmarks file.
If the bookmarks file doesn't exist QStandardPaths::locate won't find
it, what we need to do is obtain the writable location for the generic
data location and append the file name we expect. If the file exists,
VBookmarksManager will read it otherwise it will create a new one
populated with the default items.