Rename file manager bookmarks file.
It used to be tracker/bookmarks.xbel but the file manager was
renamed from Tracker to Swordfish.  I could call it swordfish/bookmarks.xbel
but filemanagement/bookmarks.xbel is a better name because the file places
model and its bookmarks will not be used only by the file manager, in fact
the same model will be used by the open/save dialogs.