Fix VSettings reading and writing and remove some complexity.
DConf < 0.13.1 has a weird API, if a key is not set we can't get
its value with dconf_client_read(), we have to use dconf_client_read_no_default()
and then if a value is not found dconf_client_read_default()
hoping for the schema to have a default value.
This seems to fix issue #22 but more tests need to be done.

With this commit I've also simplified the code by removing the
GSettings wrapper which was complicated. VSettings uses DConf much
like a configuration storage system. We might add more backends
in the future and roll our own schema definition and maybe a
schema processor that generates C++ code to read and write the
settings. With our own schema we won't have to convert GVariant
to QVariant and we'll get rid of the ugly code in VSettings::setValue()
which now reads the current value of the key just to obtain the
correct GVariant type for the QVariant conversion.